Move test_algs to docs dir, mv Plots to docs env
